Add author id and upvotes to suites
- Add Author id and Upvote count columns to db
- Update bootstrap data to match new schema
- Update /suite endpoints to check for ownership with token
- Add sys level endpoint to vote on suites
- Fixed a bug in Utils that checked incorrectly for admin roley

+++ backend/test-runner/src/main/java/com/unityTest/testrunner/restImpl/SuiteController.java
@@ -3,16 +3,19 @@ package com.unityTest.testrunner.restImpl;
 import com.unityTest.testrunner.entity.Suite;
 import com.unityTest.testrunner.entity.SuiteFile;
 import com.unityTest.testrunner.models.PLanguage;
+import com.unityTest.testrunner.models.VoteAction;
 import com.unityTest.testrunner.models.page.SuitePage;
 import com.unityTest.testrunner.models.response.FileInfo;
 import com.unityTest.testrunner.models.response.FileUploadEvent;
 import com.unityTest.testrunner.restApi.SuiteApi;
 import com.unityTest.testrunner.service.SuiteService;
 import com.unityTest.testrunner.utils.Utils;
+import org.keycloak.representations.AccessToken;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyEmitter;
@@ -35,18 +38,24 @@ public class SuiteController implements SuiteApi {
     private SuiteService suiteService;
 
     @Override
-    @RolesAllowed("ROLE_ADMIN")
-    public ResponseEntity<Suite> createTestSuite(@Valid Suite suite) {
+    public ResponseEntity<Suite> createTestSuite(Principal principal, @Valid Suite suite) {
+        String authorId = Utils.getAuthToken(principal).getSubject();       // Get author id from token
+        suite.setAuthorId(authorId);        // Set author id
+        suite.setUpvotes(0);                // Set upvote count to 0
+
         return new ResponseEntity<>(suiteService.createSuite(suite), HttpStatus.CREATED);
     }
 
-    // TODO ADD ACCESS CONTROL TO ENDPOINT
     @Override
     public ResponseEntity<FileUploadEvent> setTestSuiteFile(Principal principal, Integer suiteId, MultipartFile file) throws IOException {
-        // Get author id
-        String authorId = Utils.getAuthToken(principal).getSubject();
+        AccessToken token = Utils.getAuthToken(principal);      // Get request token
+
+        // Verify that the user uploading the file is the same as created the suite
+        Suite suite = suiteService.getSuiteById(suiteId);
+        if(!Utils.isAuthorOrAdmin(token, suite.getAuthorId())) throw new AccessDeniedException("Access Denied");
+
         // Save file uploaded to repository
-        SuiteFile suiteFile = suiteService.setSuiteTestFile(suiteId, authorId, file);
+        SuiteFile suiteFile = suiteService.setSuiteTestFile(suiteId, token.getSubject(), file);
         // Create confirmation upload event
         FileUploadEvent resp = new FileUploadEvent();
         resp.addFile(suiteFile.getFileName(), suiteFile.getFileSize());
@@ -73,9 +82,27 @@ public class SuiteController implements SuiteApi {
     }
 
     @Override
-    @RolesAllowed("ROLE_ADMIN")
-    public void deleteTestSuite(Integer suiteId) {
+    public void deleteTestSuite(Principal principal, Integer suiteId) {
+        AccessToken token = Utils.getAuthToken(principal);              // Get request token
+        Suite suiteToDelete = suiteService.getSuiteById(suiteId);       // Find suite to delete
+
+        // Check that user deleting is author or admin
+        if(!Utils.isAuthorOrAdmin(token, suiteToDelete.getAuthorId())) throw new AccessDeniedException("Access Denied");
+        // If allowed, delete the test suite
         suiteService.deleteSuite(suiteId);
+    }
+
+    @Override
+    @RolesAllowed("ROLE_SYS")
+    public void voteOnTestSuite(Integer suiteId, String action) {
+        // Convert action to VoteAction
+        VoteAction voteAction;
+        try {
+            voteAction = VoteAction.valueOf(action);
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException("Not one of accepted values for vote action");
+        }
+        suiteService.updateSuiteUpvotes(suiteId, voteAction);
     }
 
     @Override
